[NO-JIRA] Migrate storybook config
