+++ .storybook/main.js
@@ -1,7 +1,7 @@
 /*
  * Backpack - Skyscanner's Design System
  *
- * Copyright 2018 Skyscanner Ltd
+ * Copyright 2016-2019 Skyscanner Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,12 @@
  * limitations under the License.
  */
 
-// TODO: addon-a11y breaks the calendar example
-import '@storybook/addon-a11y/register';
-import '@storybook/addon-actions/register';
-import '@storybook/addon-knobs/register';
-import '@storybook/addon-viewport/register';
+module.exports = {
+  stories: ['../packages/*/stories.js'],
+  addons: [
+    '@storybook/addon-a11y',
+    '@storybook/addon-actions',
+    '@storybook/addon-knobs',
+    '@storybook/addon-viewport',
+  ],
+};
